revert moving progress.isatty() to osutil.Isatty()
